+++ Changelog.md
@@ -1,3 +1,9 @@
+# vNext
+
+Improvements:
+
+* Step completion should use word containment for filtering the list (Issue #28)
+
 # v1.3.0 - 2019-06-20
 
 This is the first open-source release from GitHub: https://github.com/specsolutions/deveroom-visualstudio