+++ Changelog.md
@@ -8,7 +8,9 @@ New features:
 * Deveroom configuration file support
   * see http://speclink.me/deveroomconfig for details
 * Allow opening artifacts related to scenarios in browser, based on configured 
-  tags that match to a specific pattern.
+  tags that match to a specific pattern (automatically recognizes 
+  [SpecSync for Azure DevOps](https://www.specsolutions.eu/services/specsync/) tags
+  and configurable for others).
   * see http://speclink.me/deveroomtraceability for details
 
 Improvements: